Moving post_save handler to receivers.py
Fixes #8
