+++ pinax/documents/apps.py
@@ -1,3 +1,5 @@
+import importlib
+
 from django.apps import AppConfig as BaseAppConfig
 from django.utils.translation import ugettext_lazy as _
 
@@ -6,3 +8,6 @@ class AppConfig(BaseAppConfig):
 
     name = "pinax.documents"
     verbose_name = _("Pinax Documents")
+
+    def ready(self):
+        importlib.import_module("pinax.documents.receivers")
