+++ pinax/documents/models.py
@@ -5,8 +5,6 @@ import uuid
 
 from django.core.urlresolvers import reverse
 from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from django.utils import timezone
 
 from django.contrib.auth import get_user_model
@@ -314,10 +312,3 @@ class UserStorage(models.Model):
         if p >= 90 and p <= 100:
             return "danger"
         raise ValueError("percentage out of range")
-
-
-@receiver(post_save, sender=settings.AUTH_USER_MODEL)
-def ensure_userstorage(sender, **kwargs):
-    if kwargs["created"]:
-        user = kwargs["instance"]
-        UserStorage.objects.create(user=user, bytes_total=(1024 * 1024 * 50))
